Improve wheel game adding new commands

+++ src/main/java/me/astri/casino/casino/BankAccount.java
@@ -1,13 +1,13 @@
 package me.astri.casino.casino;
 
 public class BankAccount {
-    private int balance;
+    private long balance;
 
     public BankAccount() {
         this.balance = 500;
     }
 
-    public boolean withdraw(int value) {
+    public boolean withdraw(long value) {
         if(value <= 0) throw new IllegalArgumentException("Value must be positive");
         if(value > balance)
             return false;
@@ -15,7 +15,7 @@ public class BankAccount {
         return true;
     }
 
-    public int getBalance() {
+    public long getBalance() {
         return this.balance;
     }
 
@@ -23,11 +23,11 @@ public class BankAccount {
         this.balance=0;
     }
 
-    public void set(int value) {
+    public void set(long value) {
         this.balance = value;
     }
 
-    public void add(int value) {
+    public void add(long value) {
         if(value <= 0) throw new IllegalArgumentException("Value must be positive");
         balance+=value;
     }
